PSQ_CreateTestpulseEpochs: Enhance error checks
Currently we only check the total number of test pulses after
PSQ_CreateTestpulseEpochs returns.

This results in code duplication on the caller side, and also does not
catch cases with wrong layout properly, as the user mostly gets an
assertion first.

We can also be nicer and just check if we have enough epochs.

Let's fix that.