ED_FindIndizesAndRedimension: Use double type for indizes wave
We have used the unsigned integer type since b9161ced (Rework labnotebook
numeric and text code, 2015-03-20) but also set the wave contents to NaN.
IP9 converts a NaN value to 4294967295 == 2^32-1.

This actually works as an initial "this-value-is-too-large-for-an-index"
but still leaves the reader puzzled.

Let's just use a double wave. We are writing double values into it anyway.