WBP_CreateWaveBuilderPanel: Clear all left over panel data on creation
Igor Pro currently does not store dimension labels for non-existing
dimensions. This means that column labels don't survive experiment saving
and loading if the row count is zero.

For the Wavebuilder this affects WBP_GetAnalysisParamGUIListWave and
friends.

The easiest solution is to throw away any existing waves on panel creation.