SF: fix SF_GetArgumentTop for zero sized input
SF_GetArgumentTop was not aware that the JSON data can be a zero sized array,
that is parsed to [null] by the FormulaParser.
SF_GetArgumentTop sent this array as is to the Executor that generates an
SF_ASSERT when null elements are encountered.

The correct solution is here to return a zero sized array to the operation
code.