EP_GetEpochs: Add support for old epoch tag formats
The shortnames in epoch tags are not present for all epochs wave versions. Older
formats, see the comment in GetEpochsWave, which don't have shortnames
can currently not be used for querying epoch information.

This is now fixed. We don't require for the full tag to be present, but
only one entry from the semicolon list is enough. Also supports regexps.