ExtractLogbookSlice: Stop changing existing slice waves
Since forever we have tried to update the slice waves on the fly if the
underlying logbook wave changed.

But the current approach breaks plots which display the existing slice in
a graph, as the slice is thrown away and recreated from scratch.

A straightforward fix would now not throw the slice away and only update
it.

But actually this function should not update the slice at all, as that is
a potentially expensive operation. And you only want to do that if you
have to.

So let's just resize the slices by the callers who write into the logbook
waves, all current callers do that, and just create the initial slice.

This also makes the function as faster. This is also relevant as it is
called from ED_createTextNotes/ED_createWaveNotes.

While at it, we also drop copying the wave note. The wave note is
confusing to have duplicated.