bugfix: Add support for free datafolders in GetListOfObjects
GetListOfObjects used datafolder string paths to build the recursive path.
This did not work if the base data folder is a free data folder.

The datafolder reference is now built using DFREF directly.

Also when using fullPath == 1 the prefix for the waves included the meaningless
base folder name when the base folder was a free data folder.
Now the full path relativ to the free data folders root is returned.