DAP_UpdateChanAmpAssignPanel: Remove call to DAP_GetNiceAmplifierChannelList
We have added that call in 35528eff (Amplifier <-> Headstage connection:
Don't preselect NONE, 2016-04-07) in order to handle the case that no
amplifiers are available.

But that is since 63023685 (Dynamically generate the amplifier popup menu
entries, 2016-03-18) always true as we return NONE plus an unselectable
"MC not available" entry.

So let's ditch that call.