PSQ_Chirp: Rework chirp cycle code completely
Even with the recent fixes to PSQ_CR_GetXPosFromCycles the current code is
buggy as FindLevels does find too many zero crossings for certain
stimsets.

Instead of now tuning an inherently heuristic approach we now use the
newly available epoch information for the full cycles in chirp stimsets.
As this uses an analytical solution it is by itself more robust.

We also add a new user epoch which spans over all `NumberOfChirpCycles`,
as this allows the user to see which range was evaluated.

The labnotebook entry PSQ_FMT_LBN_CR_CYCLES was removed as the user epoch
info replaces it. The labnotebook entry PSQ_FMT_LBN_CR_STIMSET_QC is new.
This is needed if the user has an unsuitable stimset or requested too many
cycles.