tools/clean_mies_installation.sh: Remove old MIES installations
Since forever we are just installing the new MIES version on top of the
old one. This can create problems if two XOPs from different branches
export the same operations/functions. And this is the case since 69a22f1b
(XOPs: Update TUF XOP to the released version, 2022-07-25) got merged.

The '/CIS' installation mode does not uninstall the already existing
version, so we need to do that ourselves here.