AD_FillWaves: Handle sweeps without analysis functions better
If the stimset does not have an analysis function attached, we assumed
until now that it is never ongoing. But that is wrong as these stimsets
can also have multiple sweeps.

We now check if the SCI is different than the SCI of the last acquired
sweep and that data acquisition is ongoing. This works, but needs a final
update after DAQ.

Bug introduced in 5cacbd0f (AD_FillWaves: Handle prematurely stopped
sweeps better, 2021-10-22).