AD_GetSquarePulseFailMsg: Handle broken PSQ_SquarePulse
We have reports [1] about an yet-to-be fixed MIES bug where the
PRE_SET_EVENT is not called and only one sweep is acquired.

This results in the step size labnotebook entry being not present.

Let's handle that case with a BUG message so that people can still look at
that data.