DAP_CheckSettings: Handle updated stimsets via ST_SetStimsetParameter better
When users update the stimsets via ST_SetStimsetParameter there is no hook
involved which would notify MIES that it needs to update its internal
state.

This is relevant for "Total number of steps" as that influences how DAQ is
run.

By updating the DAQ controls unconditionally here, we catch the latest
changes. DAP_UpdateDAQControls also does AFM_UpdateAnalysisFunctionWave so
we can ditch that.

It was also an oversight to ever call AFM_UpdateAnalysisFunctionWave for
TP.