ShowTraceInfoTags: Call it always
When opening a saved experiment we don't re-execute the BSP_InitPanel code.
And the only place which we know is always executed are the Igor hooks.

The hooks IgorStartOrNewHook/IgorBeforeNewHook don't work, so we use
AfterCompiledHook.

This complements the fix from da7cc293 (ShowTraceInfoTags: Call it only
once per Databrowser, 2022-05-10).