SF: Adapt logic how x-waves are mapped to y-waves
if there is a single x data wave and its size equals the number of y data waves
and the current mapped y data wave (number n) has exactly one data point,
then the n'th point of the x data wave is used as x-wave for the y-wave.

if there is a single x data wave and the other conditions are not met, then
the x data wave is mapped to all y data waves.

if there are as many x data waves as y data waves then each y data wave gets a
x data wave mapped.