TP_TSAnalysis: Change the way we calculate the instantaneous resistance
Since forever we have searched TP_EVAL_POINT_OFFSET points after the TP
pulse onset for a minimum/maximum and have then averaged +1/-1 one point
around the found extrema.

The usage of TP_EVAL_POINT_OFFSET is not correct, as the number of points
should decrease with decreasing sampling frequency so that the time which
TP_EVAL_POINT_OFFSET refers to stays constant. This is now fixed and we
use evalOffsetPointsCorrected always instead.

In addition we also use just the extrema for calculating the value. The
dropped Duplicate call is an obvious optimization.