HW_WriteDeviceInfo: Fix TTL calculation
The initial calculation for the number of TTL channels in 3fdbb60c
(DAEphys: Add a wave with device information, 2019-09-06) was fixed in
2da6ee1b (HW_WriteDeviceInfo: Fix TTL count for ITC hardware, 2021-03-17).

That fix did not take into account that in MIES we only support the front
TTLs on the ITC hardware.

The number of TTL channels also directly influences what the user can
select, as unavailable channels are greyed out by
DAP_AdaptPanelForDeviceSpecifics after locking.