SF: Fix RTE at check for dataUnit with no data waves in SF_GatherFormulaResults
SF_GatherFormulaResults has a logic to determine if data units should be shown
in the final formula plot or not. The string that is used to check for the
presence of valid data units was not handled correctly.
For the special case that no data wave are present, the IsEmpty test after
the skipped loop over all data waves generated an RTE.

introduced in 74d4d3c8a5b530390e25715ca28c3609168e4a86