DQM_FIFOMonitor: Handle FIFO thread timeouts gracefully
In 0c2a541 (TP MD: Avoid endless loop in case the monitoring thread dies,
2018-01-12) already considered and fixed the case that TP MD does not
finish properly when the monitoring thread dies. Exactly the same case can
happen with DAQ, but that was not considered back then.

We now try to restart data acqisition up to 10 times when we hit a FIFO
timeout situation.

This is also tested by supplying a fake monitor thread which always claims
to timeout.

As this is purely ITC hardware issue, the test is not being run for NI
hardware.