DAP_CheckProc_Channel_All: Don't assert out
Since 30500cb5 (DAP_LockDevice: Adapt DaEphys panel depending on device
capabilities, 2021-03-17) we grey out channel controls which are not
usable with the locked hardware.

And since e29c778d (PGC_SetAndActivateControl: Make setting a disabled
control an error, 2021-09-30) setting a disabled control is an error.

The fix is to skip disabled controls.