ITC hardware: Improve retry logic
Since forever we have retried the last ITC XOP operation when the device
in question was locked to another thread.

As we have reports from Tim about ITC1600 hardware failing quite often due
to DSP timeout errors we know also retry in this case for at most 10
times.

And to make the code nicer we wrap that code in HW_ITC_ShouldContinue.