ParseISO8601TimeStamp: Accept also standard timezone format
Since forever we required that we have a colon (:) between the timezone
hours and minutes. Worse if the colon was missing we treated all four
digits as hours.

By limiting the timezone hours and minutes to 2 digits each and making the
colon optional we can parse that format as well.