Split DoFFT function, rename it to DoPowerSpectrum, move to MIES_Utiltities
The original DoFFT function applied FFT for a specific given column of a 2D
input wave. Then it applied magsqr to calculate the power spectrum from it.

The core FFT calculation is factored out into its own function with support
for applying a window function.

The former DoFFT function was renamed to DoPowerSpectrum as it is what it
actually does.

Both functions were moved to MIES_Utilities.ipf