bugfix in IgorBeforeQuitHook ignore error on SaveExperiment
if MIES is included in Igor Pro and the user has an experiment loaded from a
read-only location then the SaveExperiment call in the Hook creates a
runtime-error that could pop up the Debugger.

Thus, if Igor Pro is quitting and the experiment can not be saved, we ignore
the error from the save attempt.