Databrowser/SweepBrowser: Add mode property
Until now we treated all databrowser/sweepbrowsers equally. For a future
application we want to use a databrowser for automation purposes.

For this we introduce a new mode which is stored in the window user data,
can be choosen when opened, and also found.

The settings history and the sweep controls subwindows are disabled in
the automation mode.

The call to DB_FindAllDataBrowser in SWS_AfterSweepDataChangeHook needs
the ALL variant as we want to update all databrowsers.