Analysis Functions: Allow to check unknown parameter names
Until know we only checked required and optional parameters via the
_CheckParam function.

For dynamic parameter names where the names are not known beforehand and
can thus not be listed in _GetParams this is a serious limitation.

Raise that limitation by passing know all parameter names,
required/optional/present into the _CheckParams function.