AfterCompiledHook: Use more threads
MultiThreadingControl allows to adjust when automatic threads are created
for calculations.

The default (mode = 1) does use automatic threads only in the main thread.
From other projects we know that this can be a limitation which is
difficult to realize and thus lift.

Let's choose mode == 4 which uses automatic threads also in explicit/preemptive
user threads.