tools/initial-repo-config.sh: Adjust the fetch refspec
Since e0d04315 (.gitmodules: Clone the submodules shallowly, 2021-04-08)
we clone the submodules shallowly to speedup CI.

Cloning a repository shallowly creates the fetch refspec

+refs/heads/<defaultbranch>:refs/remotes/origin/<defaultbranch>

And due to this refspec it is not possible to fetch other branches.

By adjusting the refspec of all submodules to the default one, we allow
fetching other branches as well. This is not detrimental for CI as the
shallow clone of the submodules is already done once this script is run.