PA_AutomaticTimeAlignment: Create reproducible results
Since the introducation of MultiThreadingControl setMode = 4, we also use
automatic multithreading in threadsafe functions.

This makes the execution of WaveStats non-reproducible if we are searching
for maximum values and we have multiple matches.

Let's use serial execution for this particular part to allow for
reproducible results.

It is not a problem for TP_TSAnalysis as data is too small. Other
cases where WaveStats is used in threadsafe functions are also good as
they don't evaluate the resulting locations.