Asyn: Fix WARN(1) in tests for orderless execution
Originally the WARN(1) was intended to generate a output message if the
test for orderless execution resulted in an ordered workload output.
(There is a chance that this happens)

However, only WARN(0) ever creates an output message.

This is fixed with this commit.

since 601010fcb05670b1689bb2b317ec0382284d0b93