WB_TrigCalculateInflectionPoints: Fix calculation
The calculation from 80dfc18f (WB_TrigSegment: Use new-style function parameter declaration, 2022-07-18) is wrong.
Redoing the math properly allows us to define a lower and upper bound for
the index of the matches.

We are now also much more confident about the solution, therefore the BUG
messages were upgraded to an assertion.