Rename XXX_WINDOW_TITLE constants to XXX_WINDOW_NAME
We do use the title constants as if these are always the same as the name.
Initially the window title it the same as the name, but can be changed be
either renaming the window or setting a new title.

In all cases we are using the title constants as the window names, so
let's rename them.