SF: Change TP annotation to be aware of invalid sweep numbers
When the tp operation is used in debug mode it displays fitted data on top
of sweep data. For the fitted data no sweep number is set as it is no true
sweep data.
In that case, while the data type is SF_DATATYPE_TP the output is mixed with
waves with true sweep data and fitted sweep data on limited ranges.
The supplementary fitted data is then just annotated with TP here.