Unify and fix "Delay onset auto" calculation
The "Delay onset auto" describes the length of the inserted TP. This is
since forever calculated in two different ways. First in
DAP_UpdateOnsetDelay from the GUI values and then later on in
DC_ReturnTotalLengthIncrease again.

Both routines used slightly different rounding. This can result in an epoch
coordinates mismatch, see test EP_EpochTest15, which created a gap between
epochs of tree level 0.

We now fix that by:
- Only writing to the "Delay onset auto" GUI variable but never reading
  again. This avoids problems with the default precision of 5 digits (%g)
- Only use this GUI value for informing the user, but not for the actual
  calculation. We do use DataConfigurationResult::testPulseLength for the
  actual calculation which is the length in points of the inserted TP
  already using the DAQ sampling interval

DAP_TPGUISettingToWave now calls DAP_UpdateOnsetDelay at the end so that
it sees the new values already.

And add a test which verifies the new behaviour.