SF_FormulaPlotter: Always keep axes ranges with constant SF code
The plot preparations are now all done in SF_PreparePlotter. This allows
us to gather the cursor infos and the axes ranges correctly before redoing
the panels.

And then later one we can restore them if the SF code did not change.

This change also required to fixup some tests which were expecting the old
naming scheme. A minor change in behaviour is that now all graphs are
created beforehand, so in case some graph could not be created it is now
empty.