Tests/SF_InsertedTPVersusTP: Make test pass again
This test currently fails always on the CI machine. The reason is that
51e251fc (CONF: Add saving and restoring MCC gain values, 2022-10-20)
started to set the MCC primary gains. And that broke the test as the
resistances for AD0 in IC mode were too different.
As every test is responsible for its own preparation we now do set the
gains we require ourselves.