TS_ThreadGroupPutDFR: Make it faster
The datafolder we pass into a ThreadGroup via ThreadGroupPutDF must not be
referenced by any DFREF variable as it crashes otherwise.

We used to workaround that issue by duplicating the passed DFREF. While
that always works it is not the fastest solution.

By passing in the DFREF as pass-by-reference variable, we can clear it so
that it is not valid anymore for the caller.