bugfix: SF left empty graph windows open on sf_assert
Commit 5614e9d introduced a change where
scaling settings are kept in the sweepformula plot windows. However, it also
changed the plot window creation to happen before the specific formulas are
parsed and executed.
Originally if a formula fails parsing or execution then no plot window is
created. This should prevent having the user to handle empty plot windows
if he entered an invalid formula. The bugged commit also disabled the test
for this behavior.

This bugfix commit reenables the test and adds a plot window cleanup logic.
So the conceptional change where all plot windows are created beforehand is
kept but on a failing formula all plot windows without displayed data are
killed. A slight disadvantage to the original concept is that on an invalid
formula empty plot windows will flicker up.