bugfix for DeepCopyWaveRefWave not reducing dimensionality of waves
It is possible to give a dimension a DeepCopyWaveRefWave where
referenced data is selectively duplicated. Previously the resulting
dimension remained at size 1, whereas it should be properly reduced
if higher dimensions are all 1 or 0.

The new utility function ReduceWaveDimensionality addresses this problem.
DeepCopyWaveRefWave uses this function to properly reduce the dimensionality.