GetDeviceInfo: Rework device info gathering
In 3fdbb60c (DAEphys: Add a wave with device information, 2019-09-06) we
added support for gathering information about devices into a permanent
wave. This was reworked in 151c3035 (Rework device info handling,
2021-03-17) to handle also unregistered devices.

In that commit we also adopted the strategy to write the device info waves
once on gathering the device lists via DAP_UpdateDeviceInfoWaves.

This worked until bb2f92b1 (CONF_RestoreDAEphys: Avoid scanning device
lists, 2022-12-13) where we skipped querying the list of devices when
using a JSON configuration file. And thus we also did not query the device
info.

We now switch to query the device info on demand in GetDeviceInfoWave().
As that works for all cases.