JSONToWave: Fix it with 2D waves and only one row
We do have the sum of all dimensions labels for each dimension. But when
one of the dimensions is 1 the sum is smaller than the here-assumed
product, so this breaks.

Bug introduced in the initial implementation in b8c8b2dc
(WaveToJSON/JSONToWave: Add them with tests, 2023-01-02).

The fix is to use dimLabelsEach as 1D array without any redimensioning.