NWB_WriteStimsetTemplateWaves: Be more cautios when warning about thirdparty stimsets
Since 8b32fe7 (NWB_WriteStimsetTemplateWaves: Handle deleted stimsets
gracefully, 2019-01-08) we warn users about stimsets which could not be
recreated.

That code also complains when customWaves are in SavedStimulusSets:DA/TTL
present as it thinks that those are third party stimsets as well. This is
confusing for users and also senior developers.

Let's not warn in this case, and also add a test for that.