EnsureLargeEnoughWave rename minimumSize argument to indexShouldExist
The argument naming minimumSize was misleading as it inferred that the value
equals a desired size. The function actually checks if the index exists that is
one less that the size (when reaching the threshold).
This led in the code to cases where too big waves are requested and additional
non-required memory reallocations where done by Redimension calls.

The new naming is more precise and should prevent wrong use of the argument.