CONF: Change DAEphys save handling if window was previously restored
If a DAEphys panel configuration is saved and the panel was previously restored
the previous configuration is now taken into account.

Configuration referred by the previous file is loaded and
- for all current existing default settings the previous settings are taken over
- all previous settings that are not existing anymore are discarded
- all new settings that have no previous setting are set default values

Rig file handling:
If a previous rig file exists then the settings defined in the rig file are
removed from the DAEphys configuration file, so there are never the same settings
distributed in DAEphys and rig file.