ITC hardware: Only open and close the ITC device once for a test run
Since forever we have occasional crashes when running the tests with ITC
hardware. In 66a0eee4
(Packages/tests/HardwareAnalysisFunctions/UTF_HardwareAnalysisFunctions.ipf:
Don't run the SC_SpikeControl tests, 2023-01-09) we already removed some
of the analysis function tests wich improved the situation considerably.

A next try is now to avoid closing an reopening the ITC device over and
over again for each test. The motivation is that user in real world see a
lot less crashes than we in CI. And one of the biggest difference is that
they don't open/close an ITC device so often.