Fix: Group creation in docker
The group was created using the user id instead of the group id. This
issue could be hidden all this time because on most systems the user id
and group id are equal, which isn't the case for Github Actions.

This bug was introduced in a3cb93991 (Documentation: Undust docker
images for documentation building, 2020-02-17).