Packages/ITCXOP2/tools/Settings.xml: Add the exception without full path
Since we added support for ITC hardware on Windows 10 in e4881c00
(CHI_CheckInstallation: Add checks for ASLR on Windows 10, 2019-08-21) we
always have added ASLR exceptions for each new major Igor Version.

But now we want to have a different path for each nightly IP version we
require, and that would require us to always add a new exception whenever
we require a new nightly version.

That does not scale, so let's just use the Executable name. We don't
support IP 32bit anymore as well.