PSX_OperationStatsImpl: Avoid histogram wave explosion
When event properties with a vast range is fed into the histogram it can
happen that the resulting wave would be too large for IP64 (200GB). This
is due to the fact that we automatically determine the bins.

Now the solution here is, in this case, not to raise the wave size limit,
but to cut out data which is too far off.