LOG_AddEntry: Redesign stacktrace storage
Up to now the stacktrace was always added as a string by callers via the
keys/values additional data.

This is suboptimal as it requires callers to gather that information
themselves. Also storing the full stacktrace as string list with
\r as separator is not well readable.

We know allow callers to add the stacktrace via an optional parameter of
LOG_AddEntry. This is off by default and only enabled for all callers
which output an error/warning.

The stacktrace is also stored as json array which results in log messages
(pretty-printed for better viewing) like:

{
  "action": "report",
  "exp": "Untitled",
  "id": "52beab5",
  "msg": "abcd",
  "source": "ReportBugToLogfile",
  "stacktrace": [
    "BUG(...)#L536 [MIES_Debugging.ipf]",
    "ReportBugToLogfile(...)#L510 [MIES_Debugging.ipf]",
    "LOG_AddEntry(...)#L151 [MIES_Logging.ipf]"
  ]
}