AD_GetPerSweepFailMessage/AD_GetBaselineFailMsg: Handle user aborted sweeps better
Investigating the uploaded MIES logfiles showed that we still have
dashboard bug messages with "Unknown reason for failure" [1].

One experiment [2] file which was acquired prior to 87f9cbfa (DAQ: Add
stopping reason to the labnotebook, 2021-05-13) showed for sweep number 5
the issue.

The sweep was terminated early before the analysis function even had a
chance to look at the baseline.

For these cases we can't bug out in AD_GetBaselineFailMsg as that is an
expected case. Instead we now look for NaNs at the end of the AD sweep
data to judge if the sweep was stopped early or not.

But other experiments have PSQ_Rheobase data [3] where we used to
determine success by just looking at the empty message from
AD_GetBaselineFailMsg. For these cases we now also return the QC state
from the function so that callers can distinguish betwen passed/failed QC
and no QC result.

[1]: https://github.com/AllenInstitute/MIES/issues/1730
[2]: Sst-IRES-Cre;Ai14-554002.08.06.02.pxp
[3]: Sst-IRES-Cre;Th-P2A-FlpO;Ai65-561491.09.09.02.pxp