PS_InitCoordinates: Allow to store the coordinates of exterior subwindows
It is now possible to store the window coordinates of resizable exterior
subwindows.

This is opt-in so you have to register the exterior subwindow via
PS_InitCoordinates in addition to the main window.

As not all subwindows might want to be saved, we now have to check in
PS_StoreWindowCoordinates if storing is requested or not.

It is not clear why MoveSubWindow needs the points -> pixel conversion.