AD_FillWaves: Handle PSQ_SquarePulse data with wrong labnotebook name
For users who acquired data with PSQ_SquarePulse between 1a64ca17
(PSQ_SealEvaluation: Add it, 2022-02-22) and be830309
(CreateAnaFuncLBNKey: Add missing break after PSQ_SQUARE_PULSE,
2022-05-20) the dashboard is currently broken.

Let's be honest and add code which determines if we need to workaround
this bug or not.

CreateAnaFuncLBNKey was taught to return the effective labnotebook prefix
applying the requested workarounds.