SF: bugfix: Add epoch range clipping for SFH_GetSweepsForFormula
After
8c6e5da (EP_WriteEpochInfoIntoSweepSettings: Handle unacquired data, 2021-07-13)
we corrected the epoch range calculation that epochs are properly clipped to
the sweep range. Older data can contain epochs that have ranges beyond the
sweep range.

For that data, we needed a workaround that clips epochs or skips outlier epochs:

- epochs outside the sweep range are skipped
- epochs starting within the sweep and ending outside are clipped
before retrieving the sweep data in that range.

Note: the original old epoch data in the LBN is not touched.