SFH: Adapt range check assertions in SFH_GetSweepsForFormula
in the old check for the range start the start should be smaller than
rightx(sweep), where a valid starting point would be the last point of the
sweep. If an epoch would start there it would be zero sized anyway.
Thus, we can check if the start is smaller than lastx, which would result
in a potential non-zero sized epoch.

In the ASSERTion check for rangeEnd the additional call to rightx was
replaced by using lastx and less or equal comparison, that is identical if
rangeEnd is locked to sample point positions.