PSX_OperationStatsImpl: Rework all event handling
The current approach was to collect all events and combo keys from
different combinations in one wave and then filter it in
PSX_GetStatsResults.

This always works, except when using the xinterval recalcuation as that
will then mix events from different combinations.

To fix that we now don't combine the events before filtering but only
after.