PSX: Fix mouse movement in stats plot
In the stats plot we can have only a subset of data plotted. So when using
the keyboard to move from event A to event B we have to think in stats
data indizes (waveIndex) and not in event indizes.

For example when you only display accepted data, you want to cycle through
the accepted data only and jump over other events.