DC: Remove DataConfigurationResult element for unAssocTP CB and LNB entry
As by default there should be no TP inserted on unassociated DA channels,
no entry for that state is needed in the DataConfigurationResult structure.

TP is only inserted in the DAQDataWave if the channel is associated. This means
the head stage is not NaN.

The LNB entry representing the Checkbox state was removed and the LNB version
increased.