SF_FormulaPlotter: Allow multiple 1D plots with the same category axis
The psxStats operation uses a category axis for the stats mode. This does
currently not work combined with the `with` statement.

Make that work by allowing multiple 1D plots with the same x-category
axis.

A test was not added as the author of this commit was unable to come up
with something which would work, see [1].

[1]: https://github.com/AllenInstitute/MIES/issues/1809