CONF: Bugfix for abort on save of DAEPhys panel
The saving of the DAEphys panel aborted if the user had the "userping" entry
in the Common configuration data block. The userping entry is a JSON object,
which is not supported by the function CopySimpleJSONElement.
Thus, transferring settings from the previous JSON configuration failed with
an abort.

Solution:
Use the more generic function JSON_SyncJSON that does not have this limitations

Since:
fdbf2308 (Add userping entries to package settings, 2023-06-01)