CONF: Do not silently abort if old rig file has unique entries
On DAEPhys save, rig settings are added to the gathered window state.
These rig settings are a combination of the default settings and if
available setting from a previous rig file. The path of a previous rig file is
stored in the DAEPHys window user data.
If such old rig file JSON paths/elements existed then these settings are
removed again from the JSON for the DAEPhys window and a separate rig file is
saved.
If due to a programmer error the previous rig file contains a JSON path that
does not appear in the gathered DAEPhys JSON the code aborted silently.

For this case an ASSERTion was added to reveal such inconsistency.