Sweepformula/psx: Add rise time
The operation psx gained a new optional parameter which accepts the return
value from the `psxRiseTime` operation. This operation is just a container
for storing the lower and upper threshold.

From the thresholds we can calculate the rise time as done in
PSX_CalculateRiseTimeImpl.

The rise time can also be queried form psxStats via `Rise Time`.

The psx operation stores the rise time in a separate cache so that when
the user tweaks the rise time parameters we don't need to redo the event
analysis. The opposite is not true as we always need to redo the rise time
calculation when the events change.