DAP: bugfix - Kill clampMode wave on DAEPhys panel locking
The consistency of the GetChannelClampMode gets corrupted.

The function DAP_RemoveClampModeSettings clears unused clamp mode settings.
However, it is only called if a headstage is unchecked or the clamp mode is
switched.
For the case the DAEPhys panel is closed and reopened the clamp mode wave with
the previous setting is kept in the former and now again current device
data folder. Though, the default headstage state in the panel is unchecked.
If one was using headstage 2 and 3 before and starts using head stage 0 and 1
in the current DAEPhys panel with at least partially the same DA/AD channels
then the setup of the new DAQ config wave fails in DC_PlaceDataInDAQConfigWave.
This is due to remaining entries in the clamp mode wave from headstage 2 and 3
that are now not used. The remaining entries cause DC_GetFilteredChannelState
to create an invalid mapping because the twice appearing DAC channel is counted
twice for a single headstage (through AFH_GetHeadstageFromDAC).

Fix:
Kill clamp mode wave on locking of DAEPhys panel.