DAEPhys: Remove control dependency of check_Settings_SaveAmpSettings
Since
d4b24fcd (DA_Ephys: Make "Save amplifier settings" checkbox dependent on "Require amplifier", 2022-07-29)
the control check_Settings_SaveAmpSettings was setup as dependent control of
check_Settings_RequireAmpConn where is gets unchecked and disabled if the main
control gets unchecked.

However, if the main control is unchecked the user can still setup amplifiers
for head stages that were not saved to the LNB then.

The fix is to remove the dependency as amplifier information should always be
saved if an amplifier was enabled and check_Settings_SaveAmpSettings was
checked.