PSX: Rename dimension labels for psxEvent
The old names taken from the original code where hard to understand. The
idea of taking them was that they might contain deeper knowledge. This
has, at least for the author of this commit, not happened, so let's name
them so that they describe what they contain.

This also raises the wave version and includes an upgrade path.