PSX: Fix initial single event fit display
When starting the psx plot with a combination different than zero, the
single event fit plot was always not shown.

The reason was that the ReplaceWave call in PSX_SetCombo did not find a
singleEventFit wave in the new comboDFR and thus did not replace that
trace.

By creating the wave for all combinations in PSX_MoveWavesToDataFolders we
ensure that the wave is always present for all combinations.