PSX_Operation: Replace peakThresh parameter by numSDs
It is more natural for the user to enter the number of stdev's from the
all points histogram than the peak threshold of the deconvoluated data.

This change required to split PSX_OperationImpl in two parts as we need
the sweepDataFiltOffDeconv data from all combinations to calculate the
peak threshold.