bugfix for showing the traceinfotags
Previously the traceinfotags were not reliably activated in the
aftercompilehook.

The new approach uses the result of the menu toggle operation
DoIgorMenu/OVRD only to determine the actual state of the menu entry.
As before a graph window must be open and visible such that the "Graph" menu is
present.

Since introduction in:
599e7973 (Add ShowTraceInfoTags, 2021-07-05)