DAP_UpdateChanAmpAssignPanel: Don't use PGC_SetAndActivateControl for control setting
This reverts commit c17cefc94 (DAP: Use PGC_SetAndActivateControl in DAP_UpdateChanAmpAssignPanel, 2023-07-24).

The reason this does not work is that using PGC calls also the GUI control
procedure. But we don't want to call the GUI procedure here as we are
syncing from CAA wave to the GUI.

Found as setting pressure devices doesn't work anymore. The hardware
pressure controls and the associations controls share the same GUI control
procedures DAP_SetVarProc_CAA and DAP_PopMenuProc_CAA and thus using PGC
on associations control also triggers a pressure control update.