Archive log files on successful locking of a device
In
3692b98f (Add archiving of uploaded parts of the Log files, 2023-05-26)
archiving of log files was introduced as part of a periodical upload of the
logs, if the user had enabled that feature.
However, if the user did not enabled that feature no archiving was done.
Still the log file data was appended to exported NWB files. With large log
files this resulted in low performance as on each NWB export the log file
was reloaded and appended.
Thus, we had to move the archiving of log files to a more generic spot and
decouple it from the upload log file dependency.

Log files are archived now when a user successfully locks a hardware device
in the DAEPhys panel. Archiving is only attempted once per MIES run.

- log files are archived if they are bigger than 50 MB
- only the part older than ~1 month is archived
- removed archiving from UploadLogFiles

UploadLogFiles uploads less logfile time than the archived time threshold,
thus there is always sufficient log data left for uploading