ConvertXOPErrorCode: Fix it for the general case
We need to handle two different ways to feed in error codes. The first is
directly the runtime error code we got from IP. These are unique and for
XOPs shifted by 10000 per XOP. The second is V_flag as returned by the XOP
with /Z being present. These are the exact same values as defined in the
XOP.

The new code handles both cases correctly.