DB: Fix assertion on unchecking sep.TTL with ITC hardware
since
31e201b9 (Util: Add GUI channel number to TUD in CreateTiledChannelGraph, 2023-07-15)

When unchecking sep.TTL when using ITC hardware always the first TTL channel
of the current rack range was referred as GUI channel. But actually it should
be the first active GUI channel. Thus, if the first channel of a rack range was
not used, then the current GUI channel was set NaN. This lead to a follow up
ASSERTion when determining the correct trace color after the color index was
calculated with a NaN result from the NaN GUI channelnumber.

The fix is to use the first active GUI channel in the current rack range.

Directly related, the axis naming for composite TTL traces was wrong as it
originally referred to the hardware channel and after 31e201b9 referred to
the first GUI channel in the rack range. This was wrong as a composite
trace does not represent this channel (unless its the only one enabled).
The name is now set to "TTLC" followed by a underscore separated list of active
GUI channels, e.g. TTLC1_3 if TTL GUI channel 1 and 3 are enabled.

Add historic data test for DataBrowser Channel Checkboxes

- integrative test that checks if switching the channel checkboxes of the
  data browser (ADC, DAC, TTL and related) causes in some combination or
  transition a RTE.